Added a rating composable to show the rating of a recipe

+++ app/src/main/java/com/example/foodium/models/WorldwideRecipe.kt
@@ -31,7 +31,7 @@ data class WorldwideRecipe(
     @SerialName("recipe_name")
     val recipeName: String,
     @SerialName("recipe_rating")
-    val recipeRating: Int,
+    val recipeRating: Float,
     @SerialName("Retinol")
     val retinol: Double,
     @SerialName("Riboflavin")
